(ref): Made separate ProposalVotes component
